ai.ohgun.site/mlsservice/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
datasets>=2.14.0
icecream>=2.1.0
requests>=2.31.0
httpx>=0.25.0
tabulate>=0.9.0
openpyxl>=3.1.0
xlrd>=2.0.1
folium>=0.16.0

# 한국어 자연어 처리
konlpy>=0.6.0
JPype1>=1.6.0

# 자연어 처리
nltk>=3.8.0

# 워드클라우드 및 시각화
wordcloud>=1.9.0
matplotlib>=3.10.0
pillow>=12.0.0